[FLINK-15956] Move the gRPC/HTTP to their own packages
