+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/grpcfn/GrpcFunction.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.statefun.flink.core.jsonmodule;
+package org.apache.flink.statefun.flink.core.grpcfn;
 
 import java.util.Objects;
 import org.apache.flink.statefun.sdk.Context;
