+++ statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/grpcfn/GrpcFunctionProvider.java
@@ -15,32 +15,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.flink.statefun.flink.core.jsonmodule;
+package org.apache.flink.statefun.flink.core.grpcfn;
 
-import java.net.SocketAddress;
-import java.util.Objects;
+import java.util.Map;
 import org.apache.flink.statefun.sdk.FunctionType;
+import org.apache.flink.statefun.sdk.StatefulFunction;
+import org.apache.flink.statefun.sdk.StatefulFunctionProvider;
 
-final class GrpcFunctionSpec implements FunctionSpec {
-  private final FunctionType functionType;
-  private final SocketAddress functionAddress;
+public class GrpcFunctionProvider implements StatefulFunctionProvider {
+  private final Map<FunctionType, GrpcFunctionSpec> supportedTypes;
 
-  GrpcFunctionSpec(FunctionType functionType, SocketAddress functionAddress) {
-    this.functionType = Objects.requireNonNull(functionType);
-    this.functionAddress = Objects.requireNonNull(functionAddress);
+  public GrpcFunctionProvider(Map<FunctionType, GrpcFunctionSpec> supportedTypes) {
+    this.supportedTypes = supportedTypes;
   }
 
   @Override
-  public FunctionType functionType() {
-    return functionType;
-  }
-
-  @Override
-  public Kind kind() {
-    return Kind.GRPC;
-  }
-
-  public SocketAddress address() {
-    return functionAddress;
+  public StatefulFunction functionOfType(FunctionType type) {
+    GrpcFunctionSpec spec = supportedTypes.get(type);
+    if (spec == null) {
+      throw new IllegalArgumentException("Unsupported type " + type);
+    }
+    return new GrpcFunction(spec);
   }
 }
