--- statefun-flink/statefun-flink-core/src/main/java/org/apache/flink/statefun/flink/core/jsonmodule/HttpFunctionSpec.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.flink.statefun.flink.core.jsonmodule;
-
-import java.net.URI;
-import java.util.List;
-import java.util.Objects;
-import org.apache.flink.statefun.sdk.FunctionType;
-
-public final class HttpFunctionSpec implements FunctionSpec {
-  private final FunctionType functionType;
-  private final URI endpoint;
-  private final List<String> states;
-
-  public HttpFunctionSpec(FunctionType functionType, URI endpoint, List<String> states) {
-    this.functionType = Objects.requireNonNull(functionType);
-    this.endpoint = Objects.requireNonNull(endpoint);
-    this.states = Objects.requireNonNull(states);
-  }
-
-  @Override
-  public FunctionType functionType() {
-    return functionType;
-  }
-
-  @Override
-  public Kind kind() {
-    return Kind.HTTP;
-  }
-
-  public URI endpoint() {
-    return endpoint;
-  }
-
-  public List<String> states() {
-    return states;
-  }
-}
